🎨 Move PastGuess interface to enums.ts
