+++ heardle_server/islands/islandProps.d.ts
@@ -1,16 +1,11 @@
-import { Signal } from "@preact/signals";
+import {Signal} from "@preact/signals";
 
-import { guessResult } from "../enums.ts";
+import {PastGuess} from "../enums.ts";
 
 interface GuessInfoProps {
   max: number;
   current: Signal<number>;
   history: Signal<PastGuess[]>;
-}
-
-interface PastGuess {
-  song?: Song;
-  result: guessResult;
 }
 
 interface CheckApiResponse {
