+++ heardle_server/islands/progress-block.tsx
@@ -1,8 +1,8 @@
 import { JSX } from "preact";
 import { useSignalEffect } from "@preact/signals";
 
-import { GuessInfoProps, PastGuess } from "./islandProps.d.ts";
-import { guessResult } from "../enums.ts";
+import { GuessInfoProps } from "./islandProps.d.ts";
+import { PastGuess, guessResult } from "../enums.ts";
 import { getSubtitleForSong } from "../helpers.tsx";
 
 const GuessStatusComponent = (
